crud

+++ src/main/java/com/example/todoapi/repositories/RoleRepository.java
@@ -3,6 +3,6 @@ package com.example.todoapi.repositories;
 import com.example.todoapi.entities.RoleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<RoleEntity, Long> {
+public interface RoleRepository extends JpaRepository<RoleEntity, Integer> {
     RoleEntity findByName(String name);
 }
